+++ src/compat/dbi.h
@@ -615,7 +615,7 @@ extern "C"
 
   extern int db_execute_statement (DB_SESSION * session, int stmt, DB_QUERY_RESULT ** result);
 
-  extern INT64 db_execute_and_keep_statement (DB_SESSION * session, int stmt, DB_QUERY_RESULT ** result);
+  extern int db_execute_and_keep_statement (DB_SESSION * session, int stmt, DB_QUERY_RESULT ** result);
   extern DB_CLASS_MODIFICATION_STATUS db_has_modified_class (DB_SESSION * session, int stmt_id);
 
   extern void db_invalidate_mvcc_snapshot_before_statement (void);
